requirements.txt
transformers
torch
datasets
pillow